refactor(desktop): split embedded terminal controller into focused hooks
